requirements.txt
notion-client==2.2.1
openai
requests
python-dotenv